whiteboard_backend/requirements.txt
Django==5.1.7
djangorestframework
channels
channels_redis
djangorestframework-simplejwt
redis
